pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dispute-desk-env"
version = "0.1.0"
description = "A real-world OpenEnv-style marketplace dispute resolution environment."
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "openenv-core>=0.2.0",
  "openai>=1.99.0",
  "pydantic>=2.10.0,<3.0.0",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.2",
  "requests>=2.32.0",
  "uvicorn>=0.35.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[project.scripts]
server = "server.app:main"
dispute-desk-server = "server.app:main"
dispute-desk-baseline = "dispute_desk.baseline:main"

[tool.setuptools.packages.find]
include = ["dispute_desk*", "server"]
